requirements.txt
Django
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib